requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-RESTful
Flask-Script
flask-cors


# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn


# ----------------------
# Extra Stuff
# ----------------------
argparse
beautifulsoup4
requests
wsgiref
autopep8
marshmallow
pep8
psycopg2
PyJWT
uWSGI